Feat: Flask backend returns index.html
The flask backend returns index.html on all routes except for any specified ones. Currently I only have one route api/predict/ that recieves an image as data2
