--- api/requirements.txt
@@ -1,10 +1,0 @@
-Flask==1.0.2
-torch==0.4.1.post2
-Flask_Cors==3.0.7
-numpy==1.15.1
-Flask==1.0.2
-torch==0.4.1.post2
-numpy==1.15.1
-Flask_Cors==3.0.7
-Pillow==5.4.1
-torchvision==0.2.1
