+++ package.json
@@ -9,7 +9,7 @@
     "build:prod": "webpack -p --env production",
     "dev-server": "webpack-dev-server",
     "test": "cross-env NODE_ENV=test jest --config=jest.config.json",
-    "start": "FLASK_APP=server/server.py flask run",
+    "start": "FLASK_APP=server.py flask run",
     "heroku-postbuild": "npm run build:prod"
   },
   "dependencies": {
